Added non-perf AR/AG unit tests to TG Nightly CI
