+++ tests/nightly/tg/ccl/test_new_all_reduce.py
@@ -1,0 +1,1 @@
+tests/ttnn/unit_tests/operations/ccl/test_new_all_reduce.py
