+++ tests/ttnn/unit_tests/operations/ccl/test_ccl_async_TG_llama.py
@@ -6,25 +6,15 @@ import torch
 import pytest
 from loguru import logger
 import ttnn
-from tests.tt_eager.python_api_testing.sweep_tests.comparison_funcs import comp_equal, comp_pcc
 from models.utility_functions import skip_for_grayskull
-from tests.ttnn.unit_tests.operations.ccl.test_ccl_common import (
-    create_and_load_sub_device_manager_with_fabric_interface,
-    teardown_fabric_interface,
-    create_global_semaphore_with_same_address,
-)
 
 from tests.ttnn.unit_tests.operations.ccl.test_all_gather_TG_post_commit import (
     run_line_all_gather_on_TG_with_mesh_tensor_along_rows,
-)
-from tests.ttnn.unit_tests.operations.ccl.test_reduce_scatter_TG_nightly import (
-    run_line_reduce_scatter_on_TG_with_mesh_tensor_along_rows,
 )
 from tests.ttnn.unit_tests.operations.ccl.test_new_all_reduce import (
     run_all_reduce_impl,
 )
 from models.perf.benchmarking_utils import BenchmarkData, BenchmarkProfiler
-from models.perf.device_perf_utils import run_device_perf_detailed
 
 
 NUM_ITERATIONS = 55
@@ -222,60 +212,6 @@ def test_all_gather_tg_llama(
     )
 
 
-@pytest.mark.parametrize(
-    "ag_type, warmup_iters, perf_target_us",
-    [
-        ("sdpa", 10, 11),
-        ("binary_mult", 10, 12),
-        ("layernorm", 10, 8),
-    ],
-)
-@pytest.mark.models_device_performance_bare_metal
-def test_ag_tg_llama_perf(
-    ag_type,
-    warmup_iters,
-    perf_target_us,
-):
-    profiler = BenchmarkProfiler()
-    benchmark_data = BenchmarkData()
-    step_name = f"all_gather_{ag_type}"
-
-    subdir = "llama_ccl_perf"
-    command = (
-        f"pytest tests/ttnn/unit_tests/operations/ccl/test_ccl_async_TG_llama.py::test_all_gather_tg_llama -k {ag_type}"
-    )
-    cols = ["DEVICE KERNEL"]
-    op_name = "AllGatherAsync"
-    warmup_iters = warmup_iters * 32  # 5 iterations per device
-
-    profiler.start("run")
-    profiler.start(step_name)
-    results = run_device_perf_detailed(command, subdir, cols, op_name, has_signposts=True, warmup_iters=warmup_iters)
-    profiler.end(step_name)
-    profiler.end("run")
-
-    # Get the measured performance
-    measured_min_us = results[cols[0]]["MIN"] / 1000
-    measured_max_us = results[cols[0]]["MAX"] / 1000
-    measured_avg_us = results[cols[0]]["AVG"] / 1000
-    measured_std_us = results[cols[0]]["STD"] / 1000
-
-    logger.info(f"Measured performance: {measured_avg_us:.3f} us vs. target: {perf_target_us} us")
-
-    # Save the measurement
-    benchmark_data.add_measurement(profiler, 0, step_name, f"all_gather-{ag_type}-min-us", measured_min_us)
-    benchmark_data.add_measurement(profiler, 0, step_name, f"all_gather-{ag_type}-max-us", measured_max_us)
-    benchmark_data.add_measurement(profiler, 0, step_name, f"all_gather-{ag_type}-avg-us", measured_avg_us)
-    benchmark_data.add_measurement(profiler, 0, step_name, f"all_gather-{ag_type}-std-us", measured_std_us)
-    benchmark_data.save_partial_run_json(
-        profiler,
-        run_type=f"all_gather",
-        ml_model_name="llama70b-tg-ccl",
-    )
-
-    assert measured_avg_us < perf_target_us, f"Performance target not met: {measured_avg_us} us > {perf_target_us} us"
-
-
 @skip_for_grayskull("Requires eth connected devices to run")
 @pytest.mark.parametrize(
     "output_shape, cluster_axis, num_links, input_num_cores, output_num_cores",
@@ -350,58 +286,3 @@ def test_all_reduce_tg_llama(
         validate_all=False,
         profiler=profiler,
     )
-
-
-@pytest.mark.parametrize(
-    "ar_type, warmup_iters, perf_target_us",
-    [
-        ("ff2", 10, 29),
-        ("qkv", 10, 25),
-        ("ff1", 10, 30),
-        ("lm_head", 10, 70),
-    ],
-)
-@pytest.mark.models_device_performance_bare_metal
-def test_ar_tg_llama_perf(
-    ar_type,
-    warmup_iters,
-    perf_target_us,
-):
-    profiler = BenchmarkProfiler()
-    benchmark_data = BenchmarkData()
-    step_name = f"all_reduce_{ar_type}"
-
-    subdir = "llama_ccl_perf"
-    command = (
-        f"pytest tests/ttnn/unit_tests/operations/ccl/test_ccl_async_TG_llama.py::test_all_reduce_tg_llama -k {ar_type}"
-    )
-    cols = ["DEVICE KERNEL"]
-    op_name = "AllReduceAsync"
-    warmup_iters = warmup_iters * 32  # 5 iterations per device
-
-    profiler.start("run")
-    profiler.start(step_name)
-    results = run_device_perf_detailed(command, subdir, cols, op_name, has_signposts=True, warmup_iters=warmup_iters)
-    profiler.end(step_name)
-    profiler.end("run")
-
-    # Get the measured performance
-    measured_min_us = results[cols[0]]["MIN"] / 1000
-    measured_max_us = results[cols[0]]["MAX"] / 1000
-    measured_avg_us = results[cols[0]]["AVG"] / 1000
-    measured_std_us = results[cols[0]]["STD"] / 1000
-
-    logger.info(f"Measured performance: {measured_avg_us:.3f} us vs. target: {perf_target_us} us")
-
-    # Save the measurement
-    benchmark_data.add_measurement(profiler, 0, step_name, f"all_reduce-{ar_type}-min-us", measured_min_us)
-    benchmark_data.add_measurement(profiler, 0, step_name, f"all_reduce-{ar_type}-max-us", measured_max_us)
-    benchmark_data.add_measurement(profiler, 0, step_name, f"all_reduce-{ar_type}-avg-us", measured_avg_us)
-    benchmark_data.add_measurement(profiler, 0, step_name, f"all_reduce-{ar_type}-std-us", measured_std_us)
-    benchmark_data.save_partial_run_json(
-        profiler,
-        run_type=f"all_reduce",
-        ml_model_name="llama70b-tg-ccl",
-    )
-
-    assert measured_avg_us < perf_target_us, f"Performance target not met: {measured_avg_us} us > {perf_target_us} us"
